requirements.txt
streamlit
streamlit-ace
openai
python-dotenv
bandit